From: Bastiaan Bakker <bastiaan.bakker@lifeline.nl> Date: Mon, 21 Nov 2005 18:34:56 +0100
Attached is a small patch against r1266 that adds a 'condrestart'
command to the RedHat initscript. This enables 'service iscsi-target
condrestart', which is needed in the iscsitarget RPM to restart ietd
after an update. The patch is also included in the 0.4.13-0.1266.1 RPM,
but it would be nice to have it in the main dist.