Date: Sun, 25 Feb 2007 10:54:39 +0200 (EET) From: Juhani Rautiainen <jrauti@iki.fi> Subject: [Iscsitarget-devel] [PATCH] Revert reserve/release to spec
I know this is not liked by many but in order to get 0.4.15 out of door 
I'll revert reserve/release to spec version. I'll make patch to get MPIO 
friendly reserve/release later. I'd also like to keep the old one as 
option which you can switch on but that is the road that leads to dark 
place. In the end you have lot's of out-of-spec things that you can  
switch on and it's maintenance nightmare. So here is the patch:

Signed-off-by: jrauti@iki.fi