Subject: Updated makefile with auto patching From: "Ross S. W. Walker" <rwalker@medallion.com> Date: Wed, 21 Mar 2007 13:01:12 -0400
1) The Makefile will properly determine the version of the current target
kernel source and automatically apply any patches defined for that kernel
version at make time. It will also detect if the target kernel version has
changed and if any patches have been applied to un-apply them and if any
patches exist for the new target to apply those. It provides two targets
'patch' and 'unpatch' that are listed as dependencies to all code targets.
There is also a helper target called 'integ_check' which makes sure that if
any patches are applied they are relevant to the current target kernel and
if not to 'unpatch' the code before patching or compiling.

2) The Makefile now properly installs all relevant executables, config
files, man pages and shared documentation. It will not install on top of any
existing config files.

3) The Makefile now properly uninstalls all relevant executables, man pages
and shared documentation. It will not uninstall any existing config files.

4) The Makefile now provides a 'distclean' function that will unapply any
previously applied patches and remove any compiled executables as well as
remove any backup files or rejects 'patch' may have left behind.

5) Makefile action targets named after their respective sub-directories,
for example 'usr' for userland code, 'kernel' for kernel code, 'doc' for
documentation etc to make it easier to guess target names.

The process for adding patches to the code base is fairly straight forward
there are two Makefile macros that handle the version checking, kver_le, for
kernel version less-than-or-equal-to and kver_ge for kernel version
greater-than-or-equal-to. These macros take a kernel version as the tuple
<major>,<minor>,<revision> (patch level isn't checked yet) and returns a
'1' if true. So for example to check whether the current kernel target
version is between 2.6.14 and 2.6.18 inclusive the Makefile section
would look something like:

ifeq ($(call kver_ge,2,6,14),1)
	ifeq($(call kver_le,2,6,18),1)
		PATCHES := $(PATCHES) compat-2.6.14-2.6.18.patch
	endif
endif