Subject: Patch to compat-2.6.14-2.6.18.patch for rev105 updates From: "Ross S. W. Walker" <rwalker@medallion.com> Date: Wed, 21 Mar 2007 13:07:03 -0400
This is the same patch Arne posted a couple of weeks ago that fixes the breakage
in compat-2.6.14-2.6.18.patch due to rev105 source code changes.

It has been thoroughly tested by both Arne and myself.

If it has already been applied then please disregard, but an svn update today
shows it has not.