From: "Ross S. W. Walker" <rwalker@medallion.com> Subject: RHEL4/CentOS4 compatibility patch Date: Thu, 22 Mar 2007 10:00:23 -0400
Here is a patch that provides RHEL4/CentOS4 compatibility fixes. It
uses the new Makefile patch method and is auto-detected and
auto-applied.

This should also demonstrate how to add distro-specific patches
into the code-base.

With long-term support distros becoming more and more popular there
is going to be a real need to add support for them over time. These
patches should not require a lot of maintenance as their platforms
are static, but the primary patches (those that are listed first)
should make the best attempt to keep a static snapshot of the code
base so as not to break the later patches.

It was thoroughly tested by Arne and myself during the development of
the latest Makefile.