Subject: Fix WCE and RCD bits of the caching mode page From: Arne Redlich <agr@powerkom-dd.de> Date: Thu, 19 Jun 2008 14:04:30 +0200
This patch is a RFC, since it hard-wires block-io's cache settings to
WCE=0 and RCD=1. This is in line with the values we return for fileio
sync (WCE=0) vs. async (WCE=1) mode, but completely ignores disk or RAID
controller caches. For those it would probably be best to make these
params user configurable!?