From: "Ross S. W. Walker" <RWalker@medallion.com> Subject: Add DKMS compatibility Date: Tue, 18 Nov 2008 14:02:47 -0500
Add compatibility for Dynamic Kernel Module Support.

If DKMS is configured properly then when ever a new kernel is installed,
DKMS will automatically build and install a new iscsitarget kernel module
for that kernel. It also keeps track of multiple module/kernel versions,
built or installed.

This patch updates the Makefile and compatibility patches to use a prefix
of 1 by sticking a bogus directory prefix in front of each file being
patched.

This patch also includes a dkms.conf file in the root that has been tested
to work successfully on at least CentOS 5.2, and is preset to release 0.4.17.

Signed-off: Ross Walker <rswwalker@gmail.com>