From: Andreas Florath <sf@flonatel.org>
Worker Thread Pool

Currently the iet starts by default eight worker threads per target.
There are some limitations of this design.  The most important is,
that a lot of resources are wasted when using a larger number of
targets.

The current implementation adds the thread pool as one possibility of
using threads.  The old one is still there and it's the default one.

When the worker thread pool should be used, the kernel module must be
loaded with a parameter giving the size of the pool (worker_thread_pool_size,
the wthreads configuration of the target is ignored then).