Rework sense data handling, fixing memleaks
Refactor the sense data handling by adding a status field and a sense_buf to
struct iscsi_cmnd. This allows for a more flexible sense data creation instead
of solely relying on a "return value <-> sense key / asc / ascq" mapping as
before.

This also plugs leaks of struct tio: on several occasions a tio wasn't freed
in error paths. Instead the pointer to it was modified to point to a newly
created tio carrying sense data, leaking the original tio.