From: "Ross S. W. Walker" <RWalker@medallion.com>
iscsitarget.spec dkms/svn build options

This patch to the iscsitarget spec file adds two build options.

1) Setting %dkms will build a noarch dkms kernel module that
automatically adds itself to dkms, builds itself and installs
itself. Upon removal it removes all instances of itself. The
package maintainer will need to perform two rpmbuilds, one
on the current architecture for the userland code, and the
other with --target=noarch for the dkms kernel module.

2) Setting %svn will build IET from the current SVN code from
the repository. It sets the version number to the last checked
in revision number from the SVN and exports a copy to the
build directory.

This has been fully tested on RHEL 5.