From: "Ross S. W. Walker" <RWalker@medallion.com>
Patch move config to /etc/iet while honoring existing

This patch moves the default location for IET config files from
/etc to /etc/iet while still honoring config files in /etc.

As soon as a config file exists in /etc/iet it will be used
instead of the one in /etc.

A warning is logged if it finds config files in /etc.

- Cleaned up plain.c a little in the initialization code path

- Added tests in target_list_build for initial getsockname and
getnameinfo calls.

Compiled on RHEL5, tested on RHEL5 with MS ini.