From: Andreas Florath <sf@flonatel.org>
target-add-debug: fix debug output

There is a minor bug in the debug-output when creating a new target:
__plain_target_create() returns false (0) when it succeeds
so the construct (from plain.c:608)
   if (__plain_target_create(&tid, p, 0);
     log_debug(1, "creating target %s", p);
prints out the 'creating target' only in error cases.

The applied patch moves the positive debug output to the same level /
position (target_add()) where the failure log is done.

Signed-off-by: Andreas Florath <sf@flonatel.org>