From: "Andreas Florath" <sf@flonatel.org>
target-find-by-name: harmonize target_find_by_xxx   functions

During the re-implementation of a MaxSession feature, I recognized that
the two functions 'target_find_by_name()' and 'target_find_by_id()'
(which do mostly the same) return different things: one the target
itself the other the target id.  The target id itself can be accessed
though 'target->tid'.

The only possibility to get a target (not only the id) from a known
name is to call the 'target_find_by_name()' - which scans the list of
all targets - and with the result of this call the function
'target_find_by_id()' - which scans the list of all targets
again.

IMHO it is much easier to adapt the return value of the function
'target_find_by_name()' to directly return the target.

Signed-off-by: Andreas Florath <sf@flonatel.org>