From: Ross S. W. Walker <RWalker@medallion.com>
target.c doesn't preserve initial sockaddr

We fail to preserve the original sockaddr_storage struct for
the portal address that discovery is being performed on, so
after we cycle through the targets it uses the last struct
data which will be the one for the unspecified address.

This patch adds a second struct for holding on to the inital
address information.

Compiled and tested on RHEL5 against MS initiator.