From: Kevin Ye <kevin.ye@gmail.com>
Fix session cleanup in case of EPIPE during login

The following patch solved the incomplete session without cid problem if
SIGPIPE is received during login process. I have tested that. The patch will
remove the session without cid when the connection is closed.

Without this patch, if SGPIPE is received, a session without cid will appear
in the /proc/net/iet/session, the target will not be able to be removed, and
--op delete will cause ietd to block at line 250 in ctldev.c: server_stop():
line 250   while (proc_session_parse(ctrl_fd, &shutdown_wait_ops, NULL) < 0)
After that, the only way to recover is to kill the ietd and start it again.