From: "Ross S. W. Walker" <RWalker@medallion.com>
allow deletion of active kernel sessions (revision 2)

This patch allows the deletion of kernel space sessions with active
connections to allow for clean session reinstatement.

This revised version fixes a bug in the prior versions:

__target_del calls session_del and in the process session_del
unlocks the target to allow nthread to clean up the connections
marked closed, but it is possible that nthread could also close
another session that had connections pending close just before
the target was locked, in which case the list_for_each_entry_safe
of sessions in __target_del would be working off a stale list.
This avoids that situation by pulling off a single session from
the list head and re-testing the list to see if it's empty each
time.

Also fixed:

target_add and __target_del don't down the semaphores interruptible
which would make killing the ietd process in the case that the
iscsi kernel module bricks impossible which could slow system
restart.

Compiled and tested on RHEL5.